Refactor the drawer JS module
