+++ packages/drawer/example.html
@@ -254,7 +254,7 @@
   <script src="https://unpkg.com/wicg-inert/dist/inert.min.js"></script>
 
   <!-- Component scripts -->
-  <script src="./dist/scripts.umd.js"></script>
+  <script src="./dev/scripts.umd.js"></script>
 
   <!-- Implementation scripts -->
   <script>
