+++ packages/drawer/src/js/switchTo.js
@@ -1,46 +1,46 @@
 import { addClass, hasClass, removeClass } from '@vrembem/core/index';
 import { setInert, setOverflowHidden } from '@vrembem/core/index';
 
-export async function switchToModal(drawerKey, obj) {
+export async function switchToModal(drawerKey) {
   // Initial guards
-  const drawer = obj.getDrawer(drawerKey);
-  if (!drawer) return obj.drawerNotFound(drawerKey);
-  if (hasClass(drawer, obj.settings.classModal)) return;
+  const drawer = this.getDrawer(drawerKey);
+  if (!drawer) return this.drawerNotFound(drawerKey);
+  if (hasClass(drawer, this.settings.classModal)) return;
 
   // Enable modal state
-  addClass(drawer, obj.settings.classModal);
-  addClass(drawer, obj.settings.stateClosed);
-  removeClass(drawer, obj.settings.stateOpened);
+  addClass(drawer, this.settings.classModal);
+  addClass(drawer, this.settings.stateClosed);
+  removeClass(drawer, this.settings.stateOpened);
 
   // Dispatch custom event
-  drawer.dispatchEvent(new CustomEvent(obj.settings.customEventPrefix + 'toModal', {
+  drawer.dispatchEvent(new CustomEvent(this.settings.customEventPrefix + 'toModal', {
     bubbles: true
   }));
   return drawer;
 }
 
-export async function switchToDefault(drawerKey, obj) {
+export async function switchToDefault(drawerKey) {
   // Initial guards
-  const drawer = obj.getDrawer(drawerKey);
-  if (!drawer) return obj.drawerNotFound(drawerKey);
-  if (!hasClass(drawer, obj.settings.classModal)) return;
+  const drawer = this.getDrawer(drawerKey);
+  if (!drawer) return this.drawerNotFound(drawerKey);
+  if (!hasClass(drawer, this.settings.classModal)) return;
 
   // Tear down modal state
-  setInert(false, obj.settings.selectorInert);
-  setOverflowHidden(false, obj.settings.selectorOverflow);
-  removeClass(drawer, obj.settings.classModal);
-  obj.focusTrap.destroy();
+  setInert(false, this.settings.selectorInert);
+  setOverflowHidden(false, this.settings.selectorOverflow);
+  removeClass(drawer, this.settings.classModal);
+  this.focusTrap.destroy();
 
   // Restore drawers saved state
-  drawerKey = drawer.getAttribute(`data-${obj.settings.dataDrawer}`);
-  const drawerState = obj.state[drawerKey];
-  if (drawerState == obj.settings.stateOpened) {
-    addClass(drawer, obj.settings.stateOpened);
-    removeClass(drawer, obj.settings.stateClosed);
+  drawerKey = drawer.getAttribute(`data-${this.settings.dataDrawer}`);
+  const drawerState = this.state[drawerKey];
+  if (drawerState == this.settings.stateOpened) {
+    addClass(drawer, this.settings.stateOpened);
+    removeClass(drawer, this.settings.stateClosed);
   }
 
   // Dispatch custom event
-  drawer.dispatchEvent(new CustomEvent(obj.settings.customEventPrefix + 'toDefault', {
+  drawer.dispatchEvent(new CustomEvent(this.settings.customEventPrefix + 'toDefault', {
     bubbles: true
   }));
   return drawer;
